requirements.txt
ipython
scikit-learn<=1.4.2
pandas<=2.2.2
numpy<=1.26.4
obspy==1.4
protobuf>=3.2
scipy<=1.11.4
tensorflow-gpu<=2.10
tensorflow-addons>=0.11.2
